ABAP To The Future 3
This contains the code samples from the ABAP To The Future 3 (Fifty Shades of ABAP) book.

The package structure reflects the structure of the book - first basic DDIC objects all chapters need, then the chapters are grouped into "enablers" and "transactional" and at the lowest level there is one package per chapter.